WavePropBase

WavePropBase.ENTITIES

WavePropBase.TAGS

WavePropBase.AbstractEntity

WavePropBase.ComplexPoint2D

WavePropBase.ComplexPoint3D

WavePropBase.ElementaryEntity

WavePropBase.ElementaryEntity

WavePropBase.HyperRectangle

WavePropBase.Point1D

WavePropBase.Point2D

WavePropBase.Point3D

WavePropBase.ReferenceLine

WavePropBase.ReferenceShape

WavePropBase.ReferenceSquare

WavePropBase.ReferenceTetrahedron

WavePropBase.ReferenceTriangle

WavePropBase.SType

Base.:==

WavePropBase._new_tag

WavePropBase.abstractmethod

WavePropBase.assert_extension

WavePropBase.blockmatrix_to_matrix

WavePropBase.blockvector_to_vector

WavePropBase.boundary

WavePropBase.bounding_box

WavePropBase.enable_debug

WavePropBase.key

WavePropBase.matrix_to_blockmatrix

WavePropBase.notimplemented

WavePropBase.print_threads_info

WavePropBase.sort_by_type

WavePropBase.svector

WavePropBase.vector_to_blockvector

WavePropBase.ENTITIES — Constant

const ENTITIES

Global dictionary storing the used entity tags (the value) for a given dimension (the key).

WavePropBase.SType — Type

const SType{T} = Union{T,Type{T}}

Union type of T and its data type Type{T}. Used to simplify methods defined on singleton types where both foo(::T) and foo(::Type{T}) are required.

WavePropBase.TAGS — Constant

const TAGS::Dict{Int,Vector{Int}}

Global dictionary storing the used entity tags (the value) for a given dimension (the key).

WavePropBase.AbstractEntity — Type

abstract type AbstractEntity

Entity of geometrical nature. Identifiable throught its (dim,tag) key.

WavePropBase.ComplexPoint2D — Type

const ComplexPoint2D
const ComplexPoint2D(x1, x2)
const ComplexPoint2D(x::NTuple{2, ComplexF64})

A complex 2D point, stored in a StaticArray. ComplexPoint2D = SVector{2, ComplexF64}.

WavePropBase.ComplexPoint3D — Type

const ComplexPoint3D
const ComplexPoint3D(x1, x2, x3)
const ComplexPoint3D(x::NTuple{3, ComplexF64})

A complex 3D point, stored in a StaticArray. ComplexPoint3D = SVector{3, ComplexF64}.

WavePropBase.ElementaryEntity — Type

struct ElementaryEntity <: AbstractEntity

The most basic representation of an AbstractEntity.

Fields:

dim::UInt8: the geometrical dimension of the entity (e.g. line has dim=1, surface has dim=2, etc)

tag::Int64: an integer tag associated to the entity

boundary::Vector{ElementaryEntity}: the entities of dimension dim-1 forming the entity's boundary

WavePropBase.ElementaryEntity — Method

ElementaryEntity(dim,tag)

Construct an ElementaryEntity with an empty boundary .

WavePropBase.HyperRectangle — Type

HyperRectangle{N,T}

Hyperrectangle in N dimensions described by low_corner::SVector{N,T} and high_corner::SVector{N,T}

WavePropBase.Point1D — Type

const Point1D
const Point1D(x1)

A point in 1D space, stored in a StaticArray. Point1D = SVector{1, Float64}.

WavePropBase.Point2D — Type

const Point2D
const Point2D(x1, x2)
const Point2D(x::NTuple{2, Float64})

A point in 2D space, stored in a StaticArray. Point2D = SVector{2, Float64}.

WavePropBase.Point3D — Type

const Point3D
const Point3D(x1, x2, x3)
const Point3D(x::NTuple{3, Float64})

A point in 3D space, stored in a StaticArray. Point3D = SVector{3, Float64}.

WavePropBase.ReferenceLine — Type

struct ReferenceLine

Singleton type representing the [0,1] segment.

WavePropBase.ReferenceShape — Type

abstract type ReferenceShape{N}

Singleton types defining a referene domain/shape in ℜᴺ. Typically used to define more complex shapes as transformations mapping an ReferenceShape into some region of ℜᴹ through a given map.

See e.g. ReferenceLine or ReferenceTriangle for examples of concrete subtypes.

WavePropBase.ReferenceSquare — Type

struct ReferenceSquare

Singleton type representing the square with vertices (0,0),(0,1),(1,1),(1,0)

WavePropBase.ReferenceTetrahedron — Type

struct ReferenceTetrahedron

Singleton type representing the tetrahedron with vertices (0,0,0),(0,0,1),(0,1,0),(1,0,0)

WavePropBase.ReferenceTriangle — Type

struct ReferenceTriangle

Singleton type representing the triangle with vertices (0,0),(0,1),(1,0)

Base.:== — Method

==(Ω1::AbstractEntity,Ω2::AbstractEntity)

Two elementary entities are considered equal if their dim and tag fields match.

Notice that this implies dim and tag of an elementary entity should uniquely define it, and therefore global variables like TAGS are needed to make sure newly created AbstractEntities have a new (dim,tag) identifier.

WavePropBase._new_tag — Method

_new_tag(dim)

Generate a unique tag for an AbstractEntity of dimension dim.

The implementation consists of adding one to the maximum value of TAGS[dim]

See also: TAGS.

WavePropBase.abstractmethod — Method

abstractmethod

A method of an abstract type for which concrete subtypes are expected to provide an implementation.

WavePropBase.assert_extension — Function

assert_extension(fname,ext,[msg])

Check that fname is of extension ext. Print the message msg as an assertion error otherwise.

WavePropBase.blockmatrix_to_matrix — Method

blockmatrix_to_matrix(A::Matrix{B}) where {B<:SMatrix}

Convert a Matrix{B}, where B<:SMatrix, to the equivalent Matrix{T}, where T = eltype(B)

WavePropBase.blockvector_to_vector — Method

blockvector_to_vector(A::Vector{B}) where {B<:SVector}

Convert a Vector{B}, where B<:SVector, to the equivalent Vector{T}, where T = eltype(B)

WavePropBase.boundary — Method

boundary(ω::ElementaryEntity)

Return the vector of elementary entities making the boundary.

WavePropBase.bounding_box — Method

bounding_box(data::Vector{<:SVector})

Create a HyperRectangle to bound the points in data.

WavePropBase.enable_debug — Method

enable_debug(mname)

Activate debugging messages by setting the environment variable JULIA_DEBUG to mname.

WavePropBase.key — Method

key(ω::ElementaryEntity)

Return the unique (dim,tag) key for the elementary entity.

WavePropBase.matrix_to_blockmatrix — Method

matrix_to_blockmatrix(A::Matrix,B)

Convert a Matrix{T} to a Matrix{B}, where B<:Type{SMatrix}. The element type of B must match that of A, and the size of A must be divisible by the size of B along each dimension.

WavePropBase.notimplemented — Method

notimplemented()

Things which should probably be implemented at some point.

WavePropBase.print_threads_info — Method

print_threads_info()

Prints in console the total number of threads.

WavePropBase.sort_by_type — Method

sort_by_type(v)

Sort the elements of v into vectors vi according to their type. Return a Dict{DataType,Vector} mapping each type to a vector of that type.

Examples

v = [1,"a",3,"b"]
dict = sort_by_type(v)

WavePropBase.svector — Method

svector(f,n)

Just like Base.ntuple, but convert output to an SVector.

WavePropBase.vector_to_blockvector — Method

vector_to_blockvector(A::Vector,B)

Convert a Vector{T} to a Vector{B}, where B<:Type{SVector}. The element type of B must match that of A, and the size of A must be divisible by the size of B along each dimension.