WavePropBase

WavePropBase.Geometry

WavePropBase.Integration

WavePropBase.Utils

WavePropBase.ambient_dimension

WavePropBase.boundary

WavePropBase.boundary

WavePropBase.bounding_box

WavePropBase.center

WavePropBase.depth

WavePropBase.diameter

WavePropBase.dimension

WavePropBase.distance

WavePropBase.distance

WavePropBase.domain

WavePropBase.entities

WavePropBase.flip_normal

WavePropBase.geometric_dimension

WavePropBase.getnodes

WavePropBase.image

WavePropBase.jacobian

WavePropBase.meshgen

WavePropBase.normal

WavePropBase.parametrization

WavePropBase.radius

WavePropBase.reference_nodes

WavePropBase.return_type

WavePropBase.tag

WavePropBase.Geometry.ENTITIES

WavePropBase.Geometry.TAGS

WavePropBase.Geometry.AbstractEntity

WavePropBase.Geometry.AbstractReferenceShape

WavePropBase.Geometry.AbstractSplitter

WavePropBase.Geometry.CardinalitySplitter

WavePropBase.Geometry.ClusterTree

WavePropBase.Geometry.ClusterTree

WavePropBase.Geometry.Domain

WavePropBase.Geometry.DyadicSplitter

WavePropBase.Geometry.ElementaryEntity

WavePropBase.Geometry.ElementaryEntity

WavePropBase.Geometry.GeometricMinimalSplitter

WavePropBase.Geometry.GeometricSplitter

WavePropBase.Geometry.HyperRectangle

WavePropBase.Geometry.PointEntity

WavePropBase.Geometry.PrincipalComponentSplitter

WavePropBase.Geometry.ReferenceLine

WavePropBase.Geometry.ReferencePoint

WavePropBase.Geometry.ReferenceSquare

WavePropBase.Geometry.ReferenceTetrahedron

WavePropBase.Geometry.ReferenceTriangle

WavePropBase.Geometry._binary_split!

WavePropBase.Geometry.assertequaldim

WavePropBase.Geometry.clear_entities!

WavePropBase.Geometry.external_boundary

WavePropBase.Geometry.internal_boundary

WavePropBase.Geometry.key

WavePropBase.Geometry.new_tag

WavePropBase.Geometry.should_split

WavePropBase.Geometry.skeleton

WavePropBase.Geometry.split!

WavePropBase.ambient_dimension — Function

ambient_dimension(x)

Dimension of the ambient space where x lives. For geometrical objects this can differ from its geometric_dimension; for example a triangle in ℝ³ has ambient dimension 3 but geometric dimension 2, while a curve in ℝ³ has ambient dimension 3 but geometric dimension 1.

WavePropBase.boundary — Function

boundary(ω)

Return the boundary of ω. For a mesh element gives the d-1 dimensional elements composing its boundary, while for an entity gives the corresponding d-1 dimensional entities.

WavePropBase.bounding_box — Function

bounding_box(data)

Create an axis-aligned bounding box containing all of data.

WavePropBase.center — Function

center(Ω)

WavePropBase.diameter — Function

diameter(Ω)

Largest distance between x and y for x,y ∈ Ω.

WavePropBase.dimension — Function

dimension(space)

The length of a basis for space; i.e. the number of linearly independent elements required to span space.

WavePropBase.domain — Function

domain(f)

The domain of the function f.

WavePropBase.entities — Function

entities(Ω::Domain)
entities(M::AbstractMesh)

Return the geometrical entities composing Ω.

WavePropBase.flip_normal — Function

flip_normal(e::AbstractEntity)

Flip the orientation of the normal vector.

WavePropBase.geometric_dimension — Function

geometric_dimension(x)
geometric_dimension(Ω::Domain)

Number of degrees of freedom necessary to locally represent the geometrical object. For example, lines have geometric dimension of 1 (whether in ℝ² or in ℝ³), while surfaces have geometric dimension of 2.

When the argument is a Domain, return the largest geometric dimension encoutered.

WavePropBase.image — Function

image(f)

The image of the function f.

WavePropBase.jacobian — Function

jacobian(f,x)

Given a (possibly vector-valued) function f : 𝐑ᵐ → 𝐅ᵐ, return the m × n matrix Aᵢⱼ = ∂fᵢ/∂x̂ⱼ.

WavePropBase.meshgen — Function

meshgen(Ω::Domain)

Generate a mesh for Ω.

WavePropBase.normal — Function

normal(el,x̂)
normal(jac::SMatrix)

The unit normal vector at coordinate x̂, guaranteed to be orthogonal to all columns of jacobian(el,x).

WavePropBase.parametrization — Function

parametrization(el)

Return the underlying parametrization of el.

WavePropBase.radius — Function

radius(Ω)

Half the diameter.

WavePropBase.reference_nodes — Function

reference_nodes(::LagrangeElement{D,Np,T})

Return the reference nodes on D used for the polynomial interpolation. The function values on these nodes completely determines the interpolating polynomial.

We use the same convention as gmsh for defining the reference nodes and their order (see node ordering on gmsh documentation).

WavePropBase.return_type — Function

return_type(f)

The type returned by the function-like object f.

WavePropBase.tag — Function

tag(::AbstractEntity)

Integer tag commonly used to idetify geometrical entities.

Utils module

WavePropBase.Utils — Module

Utils

Module containing various utility functions for WaveProp.

WavePropBase.Utils.SType — Type

const SType{T} = Union{T,Type{T}}

Union type of T and its data type Type{T}. Used to simplify methods defined on singleton types where both foo(::T) and foo(::Type{T}) are required.

WavePropBase.Utils.ComplexPoint2D — Type

const ComplexPoint2D
const ComplexPoint2D(x1, x2)
const ComplexPoint2D(x::NTuple{2, ComplexF64})

A complex 2D point, stored in a StaticArray. ComplexPoint2D = SVector{2, ComplexF64}.

WavePropBase.Utils.ComplexPoint3D — Type

const ComplexPoint3D
const ComplexPoint3D(x1, x2, x3)
const ComplexPoint3D(x::NTuple{3, ComplexF64})

A complex 3D point, stored in a StaticArray. ComplexPoint3D = SVector{3, ComplexF64}.

WavePropBase.Utils.Point1D — Type

const Point1D
const Point1D(x1)

A point in 1D space, stored in a StaticArray. Point1D = SVector{1, Float64}.

WavePropBase.Utils.Point2D — Type

const Point2D
const Point2D(x1, x2)
const Point2D(x::NTuple{2, Float64})

A point in 2D space, stored in a StaticArray. Point2D = SVector{2, Float64}.

WavePropBase.Utils.Point3D — Type

const Point3D
const Point3D(x1, x2, x3)
const Point3D(x::NTuple{3, Float64})

A point in 3D space, stored in a StaticArray. Point3D = SVector{3, Float64}.

WavePropBase.Utils.abstractmethod — Method

abstractmethod

A method of an abstract type for which concrete subtypes are expected to provide an implementation.

WavePropBase.Utils.assert_extension — Function

assert_extension(fname,ext,[msg])

Check that fname is of extension ext. Print the message msg as an assertion error otherwise.

WavePropBase.Utils.blockmatrix_to_matrix — Method

blockmatrix_to_matrix(A::Matrix{B}) where {B<:SMatrix}

Convert a Matrix{B}, where B<:SMatrix, to the equivalent Matrix{T}, where T = eltype(B)

WavePropBase.Utils.blockvector_to_vector — Method

blockvector_to_vector(A::Vector{B}) where {B<:SVector}

Convert a Vector{B}, where B<:SVector, to the equivalent Vector{T}, where T = eltype(B)

WavePropBase.Utils.cart2pol — Method

cart2pol(x,y)

Map cartesian coordinates x,y to polar coordinates r,θ. The convention followed is that -π ≤ θ ≤ π.

WavePropBase.Utils.cart2sph — Method

cart2sph(x,y,z)

Map cartesian coordinates x,y,z to spherical ones r, θ, φ representing the radius, elevation, and azimuthal angle respectively. The convention followed is that 0 ≤ θ ≤ π and -π < φ ≤ π.

WavePropBase.Utils.cross_product_matrix — Method

cross_product_matrix(v)

Returns the matrix Aᵥ associated with the cross product v × ϕ so that v × ϕ = Aᵥϕ.

WavePropBase.Utils.diagonalblockmatrix_to_matrix — Method

diagonalblockmatrix_to_matrix(A::Matrix{B}) where {B<:SMatrix}

Convert a diagonal block matrix A::AbstractVector{B}, where A is the list of diagonal blocks and B<:SMatrix, to the equivalent SparseMatrixCSC{T}, where T = eltype(B).

WavePropBase.Utils.enable_debug — Method

enable_debug(mname)

Activate debugging messages.

WavePropBase.Utils.getnodes! — Method

getnodes!(filter,nodes,tree,[isterminal=true])

Like getnodes, but append valid nodes to nodes.

WavePropBase.Utils.matrix_to_blockmatrix — Method

matrix_to_blockmatrix(A::Matrix,B)

Convert a Matrix{T} to a Matrix{B}, where B<:Type{SMatrix}. The element type of B must match that of A, and the size of A must be divisible by the size of B along each dimension.

WavePropBase.Utils.notimplemented — Method

notimplemented()

Things which should probably be implemented at some point.

WavePropBase.Utils.pol2cart — Method

pol2cart(r,θ)

Map polar coordinates r,θ to cartesian coordinates x,y.

WavePropBase.Utils.print_threads_info — Method

print_threads_info()

Prints in console the total number of threads.

WavePropBase.Utils.sort_by_type — Method

sort_by_type(v)

Sort the elements of v into vectors vi according to their type. Return a Dict{DataType,Vector} mapping each type to a vector of that type.

Examples

v = [1,"a",3,"b"]
dict = sort_by_type(v)

WavePropBase.Utils.sph2cart — Method

sph2cart(x,y,z)

Map spherical coordinates r,θ,φ representing the radius, elevation, and azimuthal angle respectively, to cartesian ones x, y, z .

WavePropBase.Utils.svector — Method

svector(f,n)

Just like Base.ntuple, but convert output to an SVector.

WavePropBase.Utils.vector_to_blockvector — Method

vector_to_blockvector(A::Vector,B)

Convert a Vector{T} to a Vector{B}, where B<:Type{SVector}. The element type of B must match that of A, and the size of A must be divisible by the size of B along each dimension.

WavePropBase.depth — Function

depth(node,acc=0)

Recursive function to compute the depth of node in a a tree-like structure. Require the method getparent(node) to be implemented. Overload this function if your structure has a more efficient way to compute depth (e.g. if it stores it in a field).

WavePropBase.getnodes — Function

getnodes(filter,tree,[isterminal=true])

Return all the nodes of tree satisfying filter(node)::Bool. If isterminal, do not consider childrens of a node for which filter(block)==true.

WavePropBase.Utils.@interface — Macro

@interface f [n=1]

Declare that the function f is an interface function. The call f(args...) resolves to M.f(args...) where M is parent module of the args[n] object.

The somewhat contrived example below illustrates how this can be used to have a generic method defined in module A applied to a type defined on module B which is independent of A but which implements the interface function f:

module A
    using WavePropBase.Utils
    Utils.@interface foo
    # a method which works on any type `x` implementing the `foo` function
    do_work(x) = 2*foo(x)
end

module B
    struct Foo end
    foo(x::Foo) = 1
end

using .A
using .B
foo = B.Foo()
A.do_work(foo)

# output

2

Note that if in the example above module A implements a generic version of foo, the call A.do_work(foo) would use that method instead based on the dispatch rules.

Geometry module

WavePropBase.Geometry — Module

module Geometry

Module defining basic geometrical concepts.

WavePropBase.Geometry.ENTITIES — Constant

const ENTITIES

Global dictionary storing the used entity tags (the value) for a given dimension (the key).

WavePropBase.Geometry.TAGS — Constant

const TAGS::Dict{Int,Vector{Int}}

Global dictionary storing the used entity tags (the value) for a given dimension (the key).

WavePropBase.Geometry.AbstractEntity — Type

abstract type AbstractEntity

Entity of geometrical nature. Identifiable throught its (dim,tag) key.

WavePropBase.Geometry.AbstractReferenceShape — Type

abstract type AbstractReferenceShape{N}

A reference domain/shape in ℜᴺ.

Used mostly for defining more complex shapes as transformations mapping an AbstractReferenceShape into some region of ℜᴹ.

See e.g. ReferenceLine or ReferenceTriangle for some examples of concrete subtypes.

WavePropBase.Geometry.AbstractSplitter — Type

abstract type AbstractSplitter

An AbstractSplitter is used to split a ClusterTree. The interface requires the following methods:

should_split(clt,splitter) : return a Bool determining if the ClusterTree should be further divided

split!(clt,splitter) : perform the splitting of the ClusterTree handling the necessary data sorting.

See GeometricSplitter for an example of an implementation.

WavePropBase.Geometry.CardinalitySplitter — Type

struct CardinalitySplitter <: AbstractSplitter

Used to split a ClusterTree along the largest dimension if length(tree)>nmax. The split is performed so the data is evenly distributed amongst all children.

WavePropBase.Geometry.ClusterTree — Type

mutable struct ClusterTree{N,T,D}

Tree structure used to hierarchically sort points in N dimensions.

Each node in the tree contains a bounding_box enclosing the points with indices loc2glob[loc_idxs].

WavePropBase.Geometry.ClusterTree — Method

ClusterTree(points,splitter)
ClusterTree{D}(points,splitter)

Construct a ClusterTree from the given points using the splitting strategy encoded in splitter.

WavePropBase.Geometry.Domain — Type

struct Domain

Represent a physical domain as a union of entities.

See also: AbstractEntity, ElementaryEntity.

WavePropBase.Geometry.DyadicSplitter — Type

struct DyadicSplitter

Used to split an N dimensional ClusterTree into 2^N children until at most nmax points are contained in node or the depth dmax is reached.

WavePropBase.Geometry.ElementaryEntity — Type

struct ElementaryEntity <: AbstractEntity

The most basic representation of an AbstractEntity.

Fields:

dim::UInt8: the geometrical dimension of the entity (e.g. line has dim=1, surface has dim=2, etc)

tag::Int64: an integer tag associated to the entity

boundary::Vector{AbstractEntity}: the entities of dimension dim-1 forming the entity's boundary

WavePropBase.Geometry.ElementaryEntity — Method

ElementaryEntity(dim,tag)

Construct an ElementaryEntity with an empty boundary .

WavePropBase.Geometry.GeometricMinimalSplitter — Type

struct GeometricMinimalSplitter <: AbstractSplitter

Like GeometricSplitter, but shrinks the children's containters.

WavePropBase.Geometry.GeometricSplitter — Type

struct GeometricSplitter <: AbstractSplitter

Used to split a ClusterTree in half along the largest axis.

WavePropBase.Geometry.HyperRectangle — Type

struct HyperRectangle{N,T}

Axis-aligned hyperrectangle in N dimensions given by low_corner::SVector{N,T} and high_corner::SVector{N,T}.

WavePropBase.Geometry.PointEntity — Type

PointEntity{N,T} <: AbstractEntity

Zero-dimension geometrical entity. As a subtype of [AbstractEntity],(@ref) the (dim,tag) of all created point entities get added to the global ENTITIES. Intended usage is to build higher dimensionsional entities, and not to represent regular points such as grid points.

WavePropBase.Geometry.PrincipalComponentSplitter — Type

struct PrincipalComponentSplitter <: AbstractSplitter

WavePropBase.Geometry.ReferenceLine — Type

struct ReferenceLine

Singleton type representing the [0,1] segment.

WavePropBase.Geometry.ReferencePoint — Type

struct ReferencePoint

Singleton type representing a reference zero-dimensional entitty (i.e. a point).

WavePropBase.Geometry.ReferenceSquare — Type

struct ReferenceSquare

Singleton type representing the square with vertices (0,0),(0,1),(1,1),(1,0)

WavePropBase.Geometry.ReferenceTetrahedron — Type

struct ReferenceTetrahedron

Singleton type representing the tetrahedron with vertices (0,0,0),(0,0,1),(0,1,0),(1,0,0)

WavePropBase.Geometry.ReferenceTriangle — Type

struct ReferenceTriangle

Singleton type representing the triangle with vertices (0,0),(0,1),(1,0)

Base.:== — Method

==(Ω1::AbstractEntity,Ω2::AbstractEntity)

Two entities are considered equal geometric_dimension(Ω1)==geometric_dimension(Ω2) and abs(tag(Ω1))=abs(tag(Ω2)). For entities of co-dimension one, the sign of tag(Ω) is used to determine the orientation of the normal vector.

Notice that this implies dim and tag of an elementary entity should uniquely define it (up to the sign of tag), and therefore global variables like TAGS are needed to make sure newly created AbstractEntity have a new (dim,tag) identifier.

Base.:== — Method

===(Ω1::Domain,Ω2::Domain)

Two Domains are equal if all their entities are equal (regardless of order).

Base.in — Method

in(ω::ElementaryEntity,Ω::Domain)

Check whether an ElementaryEntity belongs to a Domain by recursively checking whether it belongs to its boundary.

Base.iterate — Function

iterate(Ω::Domain)

Iterating over a domain means iterating over its entities.

Base.keys — Method

Return all tags of the elementary entities in the domain Ω corresponding to the dimension d.

Base.keys — Method

Return all tags of the elementary entities in the domain Ω corresponding to the dimensions contained in dims.

Base.length — Method

length(Ω:::Domain)

The length of a domain corresponds to the number of elementary entities that make it.

Base.split — Method

split(rec::HyperRectangle,[axis]::Int,[place])

Split a hyperrectangle in two along the axis direction at the position place. Returns a tuple with the two resulting hyperrectangles.

When no place is given, defaults to splitting in the middle of the axis.

When no axis and no place is given, defaults to splitting along the largest axis.

WavePropBase.Geometry._binary_split! — Method

_binary_split!(node::ClusterTree,dir,pos)
_binary_split!(f,node::ClusterTree)

Split a ClusterTree into two, sorting all points and data in the process.

Passing a dir and pos arguments splits the bounding_box box of node along direction dir at position pos, then sorts all points into the resulting left/right nodes.

If passed a predicate f, each point is sorted according to whether f(x) returns true (point sorted on the left node) or false (point sorted on the right node). At the end a minimal HyperRectangle containing all left/right points is created.

WavePropBase.Geometry.assertequaldim — Method

assertequaldim(Ω1::Domain,Ω2::Domain)

Check that two domains have same dimension.

If one of the domain (or both) are empty, the assertion is assumed to be true.

WavePropBase.Geometry.clear_entities! — Method

clear_entities!()

Empty the global variables used to keep track of the various entities created.

See also: ENTITIES, TAGS

WavePropBase.Geometry.external_boundary — Method

Return the external boundaries inside a domain.

WavePropBase.Geometry.internal_boundary — Method

Return the internal boundaries inside a domain.

WavePropBase.Geometry.key — Method

key(e::AbstractEntity)

The (dim,tag) pair used as a key to identify various abstract entities.

WavePropBase.Geometry.new_tag — Method

new_tag(dim)

Generate a unique tag for an AbstractEntity of dimension dim.

The implementation consists of adding one to the maximum value of TAGS[dim]

See also: TAGS.

WavePropBase.Geometry.should_split — Method

should_split(clt::ClusterTree,splitter::AbstractSplitter)

Determine whether or not a ClusterTree should be further divided.

WavePropBase.Geometry.skeleton — Method

skeleton(Ω::Domain)

Return all the boundaries of the domain, i.e. the domain's skeleton.

WavePropBase.Geometry.split! — Method

split!(clt::ClusterTree,splitter::AbstractSplitter)

Divide clt using the strategy implemented by splitter.

WavePropBase.boundary — Method

boundary(Ω)

Return a domain comprising the external boundary of Ω.

See also: external_boundary

WavePropBase.distance — Method

distance(x::SVector,r::HyperRectangle)

The (minimal) Euclidean distance between the point x and any point y ∈ r.

WavePropBase.distance — Method

distance(r1::HyperRectangle,r2::HyperRectangle)

The (minimal) Euclidean distance between a point x ∈ r1 and y ∈ r2.

Interpolation module

WavePropBase.Interpolation.AbstractElement — Type

abstract type AbstractElement{D,T}

Fixed interpolation schemes over the domain D<:AbstractReferenceShape. This means that the basis functions used for the interpolation are knonw from the type. The type parameter T determines the return_type of the interpolant.

Instances el of AbstractElement are expected to implement:

el(x̂): evaluate the interpolation scheme at the coordinate x̂ ∈ D.

jacobian(el,x̂) : evaluate the jacobian matrix of the interpolation at the coordinate x ∈ D.

Note

For performance reasons, both el(x̂) and jacobian(el,x̂) should take as input a StaticVector and output a static vector or array.

WavePropBase.Interpolation.AbstractPolynomialSpace — Type

abstract type AbstractPolynomialSpace{D}

A polynomial space over D. This is a vector space under polynomial addition and scalar multiplication.

WavePropBase.Interpolation.LagrangeElement — Type

struct LagrangeElement{D,Np,T} <: AbstractElement{D,T}

Standard element over D <: AbstractReferenceShape commonly used in finite element methods. The underlying polynomial space is Pk{D,K}, and it evaluates to vals::SVector{Np,T} on the reference_nodes of the element.

WavePropBase.Interpolation.LagrangeLine — Type

const LagrangeLine = LagrangeElement{ReferenceLine}

WavePropBase.Interpolation.LagrangePoint — Type

const LagrangePoint = LagrangeElement{ReferencePoint}

WavePropBase.Interpolation.LagrangeSquare — Type

const LagrangeSquare = LagrangeElement{ReferenceSquare}

WavePropBase.Interpolation.LagrangeTetrahedron — Type

const LagrangeTetrahedron = LagrangeElement{ReferenceTetrahedron}

WavePropBase.Interpolation.LagrangeTriangle — Type

const LagrangeTriangle = LagrangeElement{ReferenceTriangle}

WavePropBase.Interpolation.Pk — Type

struct Pk{D,K} <: AbstractPolynomialSpace{D}

The space of all polynomials over D of degree ≤K.

When D is a hypercube in d dimensions, the precise definition is Pk{D,K} = span{𝐱ᶿ : max(θ)≤ K}; when D is a d-dimensional simplex, the space is Pk{D,K} = span{𝐱ᶿ : sum(θ)≤ K}, where θ ∈ 𝐍ᵈ is a multi-index.

WavePropBase.Interpolation.Point — Type

const Point = LagrangePoint

WavePropBase.Interpolation.TensorLagInterp — Type

struct TensorLagInterp{N,Td,T}

Generic Lagrange interpolation over an N-dimensional tensor grid. The implementation uses a multidimensional generalization of the barycentric formula.

The main constructor takes an SVector{N,Vector{Td}} containig the N one-dimensional nodes and an Array{N,T} of the function vals at the tensor product grid formed by the one-dimensional nodes.

Examples:

nx = 10
ny = 12
x   = [0.5+0.5cos((2k-1)*π/2nx) for k in 1:nx] # Chebyshev nodes
y   = [0.5+0.5cos((2k-1)*π/2ny) for k in 1:ny] # Chebyshev nodes
f   = (x) -> cos(x[1]*x[2])
vals = [f((x,y)) for x in x, y in y]
p   = TensorLagInterp(SVector(x,y),vals)
p((0.1,0.2)) ≈ f((0.1,0.2))

WavePropBase.Interpolation.chebnodes — Method

chebnodes(n,a,b)

Return the n Chebyshev poins of the first kind on the interval [a,b]. If n is known statically, consider passing Val(n) instead for speed.

WavePropBase.Interpolation.degree — Method

degree(el::LagrangeElement)

The polynomial degree of the element. A LagrangeElement of degree K and domain D belongs to the space Pk{D,K}.

WavePropBase.Interpolation.lagrange_basis — Method

lagrange_basis(nodes,[sp::AbstractPolynomialSpace])

Return the set of n polynomials in sp taking the value of 1 on node i and 0 on nodes j ≂̸ i for 1 ≤ i ≤ n. For N-dimensional tensor-product nodes represented in the form of an SVector{N,Vector{T}}, the argument sp may be ommited.

Danger

It is assumed that the value of a function on nodes uniquely determine a polynomial in sp.

WavePropBase.Interpolation.monomial_basis — Function

monomial_basis(sp::Pk)

Return an NTuple containing a basis of monomials 𝐱ᶿ spanning the polynomial space Pk.

Integration module

WavePropBase.Integration — Module

module Integration

Methods for integrating over instances of AbstractReferenceShape.

Besides some standard quadrature rules used in the FEM module, also defines singular integration routines useful for (weakly) singular integrands.

WavePropBase.Integration.AbstractQuadratureRule — Type

abstract type AbstractQuadratureRule{D<:AbstractReferenceShape}

A quadrature rule for integrating a function over the domain D.

An instance q of AbstractQuadratureRule{D} is expected to implement the following methods:

q() : return the nodes x and weights w of the quadrature rule on the reference domain D. For performance reasons, the result should depend only on the type of q.

WavePropBase.Integration.AbstractQuadratureRule — Method

(q::AbstractQuadratureRule)()

Return the quadrature nodes x and weights w on the domain(q).

WavePropBase.Integration.AbstractSingularityHandler — Type

abstract type AbstractSingularityHandler{R}

Used for handling localized integrand singularities in R.

WavePropBase.Integration.Duffy — Type

struct Duffy <: AbstractSingularityHandler{RefereceTriangle}

Change of variables mapping the ReferenceSquare to the RefereceTriangle with the property that the jacobian vanishes at the (1,0) vertex of the triangle.

Useful for integrating functions with a singularity on the (1,0) edge of the reference triangle.

WavePropBase.Integration.Fejer — Type

struct Fejer{N}

N-point Fejer's first quadrature rule for integrating a function over [0,1]. Exactly integrates all polynomials of degree ≤ N-1.

WavePropBase.Integration.Gauss — Type

struct Gauss{D,N} <: AbstractQuadratureRule{D}

Tabulated N-point symmetric Gauss quadrature rule for integration over D.

WavePropBase.Integration.GaussLegendre — Type

struct GaussLegendre{N}

N-point Gauss-Legendre quadrature rule for integrating a function over [0,1]. Exactly integrates all polynomials of degree ≤ 2N-1.

WavePropBase.Integration.GaussLegendre — Method

GaussLegendre(;order)

Construct a GaussLegendre of the desired order over the [0,1] interval.

WavePropBase.Integration.IMT — Type

struct IMT{A,P} <: AbstractSingularityHandler{ReferenceLine}

One-dimensional change of variables mapping [0,1] -> [0,1] with the property that all derivatives vanish at the point x=0.

See Davis and Rabinowitz.

WavePropBase.Integration.Kress — Type

struct Kress{P} <: AbstractSingularityHandler{ReferenceLine}

Change of variables mapping [0,1] to [0,1] with the property that the first P derivatives of the transformation vanish at x=0.

WavePropBase.Integration.KressP — Type

struct KressP{P} <: AbstractSingularityHandler{ReferenceLine}

Like Kress, this change of variables maps the interval [0,1] onto itself, but the first P derivatives of the transformation vanish at both endpoints.

This change of variables can be used to periodize integrals in the following sense. Suppose we wish to compute the integral of f(x) from 0 to 1 where f is not a1-periodic function. If ϕ is an object of type KressP, then using it as a change of variables in the integration yields a similar integral from 0 to 1 (the interval 0≤0≤1 is mappend onto itself), but with integrand given by g(x) = f(ϕ(x))ϕ'(x). Since ϕ' vanishes (together with P of its derivatives), the function g(x) is now periodic (up to derivatives of order up to P) at the endpoints. Thus quadrature rules designed for periodic functions like the TrapezoidalP can be used to obtain high order convergence of g, which in turn yields a modified quadrature rule when viewed as a quadrature rule for f.

WavePropBase.Integration.SingularQuadratureRule — Type

SingularQuadratureRule{D,Q,S} <: AbstractQuadratureRule{D}

A quadrature rule over D intended to integrate functions which are singular at a known point s ∈ D.

A singular quadrature is rule is composed of a regular quadrature rule (e.g. GaussLegendre) and a AbstractSingularityHandler to transform the regular quadrature. The regular quadrature rule generates nodes and weights on the domain(sing_handler), and those are mapped into an appropriate quadrature over D = range(sing_handler) using the singularity handler.

WavePropBase.Integration.TensorProductQuadrature — Type

TensorProductQuadrature{Q}

A tensor-product of one-dimension quadrature rules. Integrates over [0,1]^d, where d=length(quad).

Examples

qx = Fejer(10)
qy = GaussLegendre(15)
q  = TensorProductQuadrature(qx,qy)

WavePropBase.Integration.TensorProductSingularityHandler — Type

struct TensorProductSingularityHandler{S} <: AbstractSingularityHandler{ReferenceSquare}

A tensor product of two one-dimensional AbstractSingularityHandlers for performing integration over the ReferenceSquare.

WavePropBase.Integration.Trapezoidal — Type

struct Trapezoidal{N} <: AbstractQuadratureRule{ReferenceLine}

N-point trapezoidal rule for integrating a function over the interval [0,1].

For periodic functions over [0,1], see TrapezoidalP.

Examples:

q    = Trapezoidal(10)
f(x) = exp(x)*cos(x)
integrate(f,q)

WavePropBase.Integration.TrapezoidalP — Type

struct TrapezoidalP{N} <: AbstractQuadratureRule{ReferenceLine}

Open trapezoidal rule. Useful for periodic functions since it does not duplicate the boundary nodes, void duplication.

WavePropBase.Integration._get_gauss_qnodes_and_qweights — Method

_get_qnodes_and_qweights(R::Type{<:AbstractReferenceShape{D}}, N) where D

Returns the N-point symmetric gaussian qnodes and qweights (x, w) for integration over R.

WavePropBase.Integration.integrate — Method

integrate(f,q::AbstractQuadrature)
integrate(f,x,w)

Integrate the function f using the quadrature rule q. This is simply sum(f.(x) .* w), where x and w are the quadrature nodes and weights, respectively.

WavePropBase.Integration.qnodes — Method

qnodes(Y)

Return the quadrature nodes associated with Y.

WavePropBase.Integration.qrule_for_reference_shape — Method

qrule_for_reference_shape(ref,order)

Given a reference shape and a desired quadrature order, return an appropiate quadrature rule.

WavePropBase.Integration.qweights — Method

qweights(Y)

Return the quadrature weights associated with Y.

WavePropBase.Integration.refine — Method

refine(q::AbstractQuadratureRule,[k=2])

Generate a similar quadrature rule, but with k-times as many quadrature nodes.

WavePropBase.Integration.singular_quadrature — Method

singular_quadrature(k,q::SingularQuadratureRule,s)

Return nodes and weights to integrate a function over domain(q) with a factored weight k.

WavePropBase.Integration.singular_weights — Method

singular_weights(k,xi,q::SingularQuadratureRule,s)

Mesh module

WavePropBase.Mesh.AbstractMesh — Type

abstract type AbstractMesh{N,T}

An abstract mesh structure in dimension N with primite data of type T (e.g. Float64 for double precision representation).

Concrete subtypes of AbstractMesh should implement ElementIterator for accessing the mesh elements, and/or NodeIterator for accesing the mesh nodes.

See also: GenericMesh, UniformCartesianMesh

WavePropBase.Mesh.ElementIterator — Type

struct ElementIterator{E,M}

Return an iterator for all elements of type E on a mesh of type M.

Besides the methods listed in the iterator iterface of Julia, some functions also require the getindex(iter,i::Int) method for accessing the i-th element directly.

WavePropBase.Mesh.GenericMesh — Type

struct GenericMesh{N,T} <: AbstractMesh{N,T}

Data structure representing a generic mesh in an ambient space of dimension N, with data of type T.

WavePropBase.Mesh.NodeIterator — Type

struct NodeIterator{M}

Iterator for all the nodes in a mesh of type M.

WavePropBase.Mesh.SubMesh — Type

struct SubMesh{N,T} <: AbstractMesh{N,T}

Create a view of a parent mesh over a given domain.

A submesh implements the interface for AbstractMesh; therefore you can iterate over elements of the submesh just like you would with a mesh.

WavePropBase.Mesh.UniformCartesianMesh — Type

struct UniformCartesianMesh{N,T} <: AbstractMesh{N,T}

An N-dimensional cartesian grid given as the tensor-product of N one-dimensional LinRange{T} grids.

Iterating over a UniformCartesianMesh generates the elements which compose the mesh; i.e. the HyperRectangle cells.

WavePropBase.Mesh.UniformCartesianMesh — Method

UniformCartesianMesh(;domain::HyperRectangle,sz::NTuple)

Construct a uniform UniformCartesianMesh with sz[d] elements along dimension d.

WavePropBase.Mesh.decompose — Method

decompose(s::AbstractReferenceShape,x,[target_shape])

Decompose an AbstractReferenceShape into LagrangeElements so that x is a vertex of the children elements.

Examples

s = ReferenceLine() el1, el2 = decompose(s,0.3) el1(1) == el2(0) == 0.3 # true

WavePropBase.Mesh.dom2elt — Method

dom2elt(m::GenericMesh,Ω,E)

Compute the element indices idxs of the elements of type E composing Ω, so that m[E][idxs] gives all the elements of type E meshing Ω.

WavePropBase.Mesh.dom2elt — Method

dom2elt(m::GenericMesh,Ω)

Return a Dict with keys being the element types of m, and values being the indices of the elements in Ω of that type.

WavePropBase.Mesh.dom2elt — Method

dom2elt(m::SubMesh,[E])

A dictionary with keys being the element types of m, and values being the element indices in the parent mesh. If a type E is given, return the values associated with that key.

IO module

WavePropBase.IO.PlotPoints — Type

struct PlotPoints

Structure used for dispatching SVector to plot recipes without type-piracy.

WavePropBase.IO.PlotTree — Type

struct PlotTree

Used to plot entire tree associated with a tree node, instead of just the node.