WavePropBase

WavePropBase.ENTITIES

WavePropBase.TAGS

WavePropBase.AbstractElement

WavePropBase.AbstractEntity

WavePropBase.AbstractMesh

WavePropBase.CartesianMesh

WavePropBase.CartesianMesh

WavePropBase.ComplexPoint2D

WavePropBase.ComplexPoint3D

WavePropBase.Domain

WavePropBase.ElementIterator

WavePropBase.ElementaryEntity

WavePropBase.ElementaryEntity

WavePropBase.GenericMesh

WavePropBase.HyperRectangle

WavePropBase.LagrangeElement

WavePropBase.LagrangeInterp

WavePropBase.LagrangeLine

WavePropBase.LagrangeRectangle

WavePropBase.LagrangeTetrahedron

WavePropBase.LagrangeTriangle

WavePropBase.PlotPoints

WavePropBase.Point1D

WavePropBase.Point2D

WavePropBase.Point3D

WavePropBase.ReferenceLine

WavePropBase.ReferenceShape

WavePropBase.ReferenceSquare

WavePropBase.ReferenceTetrahedron

WavePropBase.ReferenceTriangle

WavePropBase.SType

WavePropBase.SubMesh

Base.:==

Base.:==

Base.in

Base.iterate

Base.keys

Base.keys

Base.length

WavePropBase.abstractmethod

WavePropBase.ambient_dimension

WavePropBase.assert_extension

WavePropBase.assertequaldim

WavePropBase.blockmatrix_to_matrix

WavePropBase.blockvector_to_vector

WavePropBase.boundary

WavePropBase.bounding_box

WavePropBase.center

WavePropBase.clear_entities!

WavePropBase.degree

WavePropBase.diameter

WavePropBase.dom2elt

WavePropBase.dom2elt

WavePropBase.dom2elt

WavePropBase.domain

WavePropBase.enable_debug

WavePropBase.entities

WavePropBase.external_boundary

WavePropBase.geometric_dimension

WavePropBase.internal_boundary

WavePropBase.jacobian

WavePropBase.key

WavePropBase.matrix_to_blockmatrix

WavePropBase.new_tag

WavePropBase.normal

WavePropBase.notimplemented

WavePropBase.print_threads_info

WavePropBase.radius

WavePropBase.reference_nodes

WavePropBase.return_type

WavePropBase.skeleton

WavePropBase.sort_by_type

WavePropBase.svector

WavePropBase.tag

WavePropBase.vector_to_blockvector

WavePropBase.@interface

WavePropBase.ENTITIES — Constant

const ENTITIES

Global dictionary storing the used entity tags (the value) for a given dimension (the key).

WavePropBase.SType — Type

const SType{T} = Union{T,Type{T}}

Union type of T and its data type Type{T}. Used to simplify methods defined on singleton types where both foo(::T) and foo(::Type{T}) are required.

WavePropBase.TAGS — Constant

const TAGS::Dict{Int,Vector{Int}}

Global dictionary storing the used entity tags (the value) for a given dimension (the key).

WavePropBase.AbstractElement — Type

abstract type AbstractElement{D,T}

Abstract element given by an interpolation scheme of return_type T over the domain D<:ReferenceShape.

Instances el of AbstractElement are expected to implement:

el(x̂): evaluate the interpolation scheme at the coordinate x̂ ∈ D.

jacobian(el,x̂) : evaluate the jacobian matrix of the interpolation at the coordinate x ∈ D.

Note

For performance reasons, both el(x̂) and jacobian(el,x̂) should typically operate on static arrays/vectors.

WavePropBase.AbstractEntity — Type

abstract type AbstractEntity

Entity of geometrical nature. Identifiable throught its (dim,tag) key.

WavePropBase.AbstractMesh — Type

abstract type AbstractMesh{N,T}

An abstract parent structure in dimension N with primite data of type T (e.g. Float64 for double precision representation).

The AbstractMesh interface expects the following methods to be implemented:

keys(msh) : return a list of the element types composing the mesh.

msh[E] : return an ElementIterator for the mesh elements of type E

See also: ElementIterator

WavePropBase.CartesianMesh — Type

struct CartesianMesh{N,T} <: AbstractMesh{N,T}

An N-dimensional cartesian grid given as the tensor-product of N one-dimensional Vector{T} grids.

Iterating over a CartesianMesh generates the elements which compose the mesh; i.e. the HyperRectangle cells.

WavePropBase.CartesianMesh — Method

CartesianMesh(;domain::HyperRectangle,sz::NTuple)

Construct a uniform CartesianMesh with sz[d] elements along dimension d.

WavePropBase.ComplexPoint2D — Type

const ComplexPoint2D
const ComplexPoint2D(x1, x2)
const ComplexPoint2D(x::NTuple{2, ComplexF64})

A complex 2D point, stored in a StaticArray. ComplexPoint2D = SVector{2, ComplexF64}.

WavePropBase.ComplexPoint3D — Type

const ComplexPoint3D
const ComplexPoint3D(x1, x2, x3)
const ComplexPoint3D(x::NTuple{3, ComplexF64})

A complex 3D point, stored in a StaticArray. ComplexPoint3D = SVector{3, ComplexF64}.

WavePropBase.Domain — Type

struct Domain

Represent a physical domain as a union of entities.

See also: AbstractEntity, ElementaryEntity.

WavePropBase.ElementIterator — Type

struct ElementIterator{E,M}

Return an iterator for all elements of type E on a mesh of type M.

Besides the methods listed in the iterator iterface of Julia, some functions also require the getindex(iter,i::Int) method for accessing the i-th element directly.

WavePropBase.ElementaryEntity — Type

struct ElementaryEntity <: AbstractEntity

The most basic representation of an AbstractEntity.

Fields:

dim::UInt8: the geometrical dimension of the entity (e.g. line has dim=1, surface has dim=2, etc)

tag::Int64: an integer tag associated to the entity

boundary::Vector{ElementaryEntity}: the entities of dimension dim-1 forming the entity's boundary

WavePropBase.ElementaryEntity — Method

ElementaryEntity(dim,tag)

Construct an ElementaryEntity with an empty boundary .

WavePropBase.GenericMesh — Type

struct GenericMesh{N,T} <: AbstractMesh{N,T}

Data structure representing a generic mesh in an ambient space of dimension N, with data of type T.

WavePropBase.HyperRectangle — Type

HyperRectangle{N,T}

Hyperrectangle in N dimensions described by a low_corner::SVector{N,T} and a high_corner::SVector{N,T}

WavePropBase.LagrangeElement — Type

struct LagrangeElement{D,Np,T} <: AbstractElement{D,T}

Fields:

vals::SVector{Np,T}

WavePropBase.LagrangeInterp — Type

struct LagrangeInterp{N,Td,T}

Generic Lagrange interpolation over an N-dimensional tensor grid. The implementation uses a multidimensional generalization of the barycentric formula.

The main constructor takes an SVector{N,Vector{Td}} containig the N one-dimensional nodes and an Array{N,T} of the function vals at the tensor product grid formed by the one-dimensional nodes.

Examples:

nx = 10
ny = 12
x   = [0.5+0.5cos((2k-1)*π/2nx) for k in 1:nx] # Chebyshev nodes
y   = [0.5+0.5cos((2k-1)*π/2ny) for k in 1:ny] # Chebyshev nodes
f   = (x) -> cos(x[1]*x[2])
vals = [f((x,y)) for x in x, y in y]
p   = LagrangeInterp(SVector(x,y),vals)
p((0.1,0.2)) ≈ f((0.1,0.2))

WavePropBase.LagrangeLine — Type

const LagrangeLine = LagrangeElement{ReferenceLine}

WavePropBase.LagrangeRectangle — Type

const LagrangeRectangle = LagrangeElement{ReferenceSquare}

WavePropBase.LagrangeTetrahedron — Type

const LagrangeTetrahedron = LagrangeElement{ReferenceTetrahedron}

WavePropBase.LagrangeTriangle — Type

const LagrangeTriangle = LagrangeElement{ReferenceTriangle}

WavePropBase.PlotPoints — Type

struct PlotPoints

Structure used for dispatching plot recipes. Its use also avoids the problem of type-piracy.

WavePropBase.Point1D — Type

const Point1D
const Point1D(x1)

A point in 1D space, stored in a StaticArray. Point1D = SVector{1, Float64}.

WavePropBase.Point2D — Type

const Point2D
const Point2D(x1, x2)
const Point2D(x::NTuple{2, Float64})

A point in 2D space, stored in a StaticArray. Point2D = SVector{2, Float64}.

WavePropBase.Point3D — Type

const Point3D
const Point3D(x1, x2, x3)
const Point3D(x::NTuple{3, Float64})

A point in 3D space, stored in a StaticArray. Point3D = SVector{3, Float64}.

WavePropBase.ReferenceLine — Type

struct ReferenceLine

Singleton type representing the [0,1] segment.

WavePropBase.ReferenceShape — Type

abstract type ReferenceShape{N}

Singleton types defining a referene domain/shape in ℜᴺ. Typically used to define more complex shapes as transformations mapping an ReferenceShape into some region of ℜᴹ through a given map.

See e.g. ReferenceLine or ReferenceTriangle for examples of concrete subtypes.

WavePropBase.ReferenceSquare — Type

struct ReferenceSquare

Singleton type representing the square with vertices (0,0),(0,1),(1,1),(1,0)

WavePropBase.ReferenceTetrahedron — Type

struct ReferenceTetrahedron

Singleton type representing the tetrahedron with vertices (0,0,0),(0,0,1),(0,1,0),(1,0,0)

WavePropBase.ReferenceTriangle — Type

struct ReferenceTriangle

Singleton type representing the triangle with vertices (0,0),(0,1),(1,0)

WavePropBase.SubMesh — Type

struct SubMesh{N,T} <: AbstractMesh{N,T}

Create a view of a parent mesh over a given domain.

A submesh implements the interface for AbstractMesh; therefore you can iterate over elements of the submesh just like you would with a mesh.

Base.:== — Method

==(Ω1::AbstractEntity,Ω2::AbstractEntity)

Two elementary entities are considered equal geometric_dimension(Ω1)==geometric_dimension(Ω2) and abs(tag(Ω1))=abs(tag(Ω2)). The sign of tag(Ω) is used to determine its orientation.

Notice that this implies dim and tag of an elementary entity should uniquely define it (up to the sign of tag), and therefore global variables like TAGS are needed to make sure newly created AbstractEntities have a new (dim,tag) identifier.

Base.:== — Method

===(Ω1::Domain,Ω2::Domain)

Two Domains are equal if all their entities are equal (regardless of order).

Base.in — Method

in(ω::ElementaryEntity,Ω::Domain)

Check whether an ElementaryEntity belongs to a Domain by recursively checking whether it belongs to its boundary.

Base.iterate — Function

iterate(Ω::Domain)

Iterating over a domain means iterating over its entities.

Base.keys — Method

Return all tags of the elementary entities in the domain Ω corresponding to the dimension d.

Base.keys — Method

Return all tags of the elementary entities in the domain Ω corresponding to the dimensions contained in dims.

Base.length — Method

length(Ω:::Domain)

The length of a domain corresponds to the number of elementary entities that make it.

WavePropBase.abstractmethod — Method

abstractmethod

A method of an abstract type for which concrete subtypes are expected to provide an implementation.

WavePropBase.ambient_dimension — Function

ambient_dimension(x)

Dimension of the ambient space where x lives. For geometrical objects this can differ from its geometric_dimension; for example a triangle in ℝ³ has ambient dimension 3 but geometric dimension 2, while a curve in ℝ³ has ambient dimension 3 but geometric dimension 1.

WavePropBase.assert_extension — Function

assert_extension(fname,ext,[msg])

Check that fname is of extension ext. Print the message msg as an assertion error otherwise.

WavePropBase.assertequaldim — Method

assertequaldim(Ω1::Domain,Ω2::Domain)

Check that two domains have same dimension.

If one of the domain (or both) are empty, the assertion is assumed to be true.

WavePropBase.blockmatrix_to_matrix — Method

blockmatrix_to_matrix(A::Matrix{B}) where {B<:SMatrix}

Convert a Matrix{B}, where B<:SMatrix, to the equivalent Matrix{T}, where T = eltype(B)

WavePropBase.blockvector_to_vector — Method

blockvector_to_vector(A::Vector{B}) where {B<:SVector}

Convert a Vector{B}, where B<:SVector, to the equivalent Vector{T}, where T = eltype(B)

WavePropBase.boundary — Function

boundary(ω)

Return the boundary of ω. For a mesh element gives the d-1 dimensional elements composing its boundary, while for an entity gives the corresponding d-1 dimensional entities.

WavePropBase.bounding_box — Method

bounding_box(data)

Create a HyperRectangle to bound the points in data.

WavePropBase.center — Function

center(Ω)

WavePropBase.clear_entities! — Method

clear_entities!()

Empty the global variables used to keep track of the various entities created.

See also: ENTITIES, TAGS

WavePropBase.degree — Method

degree(el::LagrangeElement)

The degree of the underlying polynomial used to represent this type of element.

WavePropBase.diameter — Function

diameter(Ω)

Largest distance between x and y for x,y ∈ Ω.

WavePropBase.dom2elt — Method

dom2elt(m::GenericMesh,Ω,E)

Compute the element indices idxs of the elements of type E composing Ω, so that m[E][idxs] gives all the elements of type E meshing Ω.

WavePropBase.dom2elt — Method

dom2elt(m::GenericMesh,Ω)

Return a Dict with keys being the element types of m, and values being the indices of the elements in Ω of that type.

WavePropBase.dom2elt — Method

dom2elt(m::SubMesh,[E])

A dictionary with keys being the element types of m, and values being the element indices in the parent mesh. If a type E is given, return the values associated with that key.

WavePropBase.domain — Function

domain(f)

The domain of f. For elements of geometrical nature return the ReferenceShape used to represent it.

WavePropBase.enable_debug — Method

enable_debug(mname)

Activate debugging messages by setting the environment variable JULIA_DEBUG to mname.

WavePropBase.entities — Method

entities(Ω::Domain)

Return a vector of all elementary entities making up a domain.

WavePropBase.external_boundary — Method

Return the external boundaries inside a domain.

WavePropBase.geometric_dimension — Function

geometric_dimension(x)
geometric_dimension(Ω::Domain)

Number of degrees of freedom necessary to locally represent the geometrical object. For example, lines have geometric dimension of 1 (whether in ℝ² or in ℝ³), while surfaces have geometric dimension of 2.

When the argument is a Domain, return the largest geometric dimension encoutered.

WavePropBase.internal_boundary — Method

Return the internal boundaries inside a domain.

WavePropBase.jacobian — Function

jacobian(F,x̂)

The Jacobian matrix Aᵢⱼ = ∂Fᵢ/∂x̂ⱼ at the parametric coordinate x̂.

WavePropBase.key — Method

key(e::AbstractEntity)

The (dim,tag) pair used as a key to identify various abstract entities.

WavePropBase.matrix_to_blockmatrix — Method

matrix_to_blockmatrix(A::Matrix,B)

Convert a Matrix{T} to a Matrix{B}, where B<:Type{SMatrix}. The element type of B must match that of A, and the size of A must be divisible by the size of B along each dimension.

WavePropBase.new_tag — Method

new_tag(dim)

Generate a unique tag for an AbstractEntity of dimension dim.

The implementation consists of adding one to the maximum value of TAGS[dim]

See also: TAGS.

WavePropBase.normal — Function

normal(el,x̂)

The unit normal vector of el at the parametric coordinate x̂.

WavePropBase.notimplemented — Method

notimplemented()

Things which should probably be implemented at some point.

WavePropBase.print_threads_info — Method

print_threads_info()

Prints in console the total number of threads.

WavePropBase.radius — Function

radius(Ω)

Half the diameter.

WavePropBase.reference_nodes — Function

reference_nodes(::LagrangeElement{D,Np,T})

Return the reference nodes on D used for the polynomial interpolation. The function values on these nodes completely determines the interpolating polynomial.

We use the same convention as gmsh for defining the reference nodes and their order (see node ordering on gmsh documentation).

WavePropBase.return_type — Function

return_type(f)

The type returned by the function-like object f.

WavePropBase.skeleton — Method

skeleton(Ω::Domain)

Return all the boundaries of the domain, i.e. the domain's skeleton.

WavePropBase.sort_by_type — Method

sort_by_type(v)

Sort the elements of v into vectors vi according to their type. Return a Dict{DataType,Vector} mapping each type to a vector of that type.

Examples

v = [1,"a",3,"b"]
dict = sort_by_type(v)

WavePropBase.svector — Method

svector(f,n)

Just like Base.ntuple, but convert output to an SVector.

WavePropBase.tag — Function

tag(::AbstractEntity)

Integer tag used to idetify geometrical entities.

WavePropBase.vector_to_blockvector — Method

vector_to_blockvector(A::Vector,B)

Convert a Vector{T} to a Vector{B}, where B<:Type{SVector}. The element type of B must match that of A, and the size of A must be divisible by the size of B along each dimension.

WavePropBase.@interface — Macro

@interface f [n=1]

Declare that the function f is an interface function. The call f(args...) resolves to M.f(args...) where M is parent module of the args[n] object.

The somewhat contrived example below illustrates how this can be used to have a generic method defined in module A applied to a type defined on module B which is independent of A but which implements the interface function f:

module A
    using WavePropBase
    WavePropBase.@interface foo
    # a method which works on any type `x` implementing the `foo` function
    do_work(x) = 2*foo(x)
end

module B
    struct Foo end
    foo(x::Foo) = 1
end

using .A
using .B
foo = B.Foo()
A.do_work(foo)

# output

2

Note that if in the example above module A implements a generic version of foo, the call A.do_work(foo) would use that method instead based on the dispatch rules.