WavePropBase

WavePropBase.AbstractReferenceShape

WavePropBase.ComplexPoint2D

WavePropBase.ComplexPoint3D

WavePropBase.Point2D

WavePropBase.Point3D

WavePropBase.ReferenceLine

WavePropBase.ReferenceSquare

WavePropBase.ReferenceTetrahedron

WavePropBase.ReferenceTriangle

WavePropBase.SType

WavePropBase.abstractmethod

WavePropBase.assert_extension

WavePropBase.blockmatrix_to_matrix

WavePropBase.blockvector_to_vector

WavePropBase.enable_debug

WavePropBase.matrix_to_blockmatrix

WavePropBase.notimplemented

WavePropBase.print_threads_info

WavePropBase.svector

WavePropBase.vector_to_blockvector

WavePropBase.SType — Type

const SType{T}

Union type of T and its data type Type{T}. Used to simplify methods defined on singleton types where both foo(::T) and foo(::Type{T}) are required.

WavePropBase.AbstractReferenceShape — Type

abstract type AbstractReferenceShape{N}

A reference domain/shape in ℜᴺ.

Used mostly for defining more complex shapes as transformations mapping an AbstractReferenceShape into some region of ℜᴹ.

See e.g. ReferenceLine or ReferenceTriangle for some examples of concrete subtypes.

WavePropBase.ComplexPoint2D — Type

const ComplexPoint2D
const ComplexPoint2D(x1, x2)
const ComplexPoint2D(x::NTuple{2, ComplexF64})

A complex 2D point, stored in a StaticArray. ComplexPoint2D = SVector{2, ComplexF64}.

WavePropBase.ComplexPoint3D — Type

const ComplexPoint3D
const ComplexPoint3D(x1, x2, x3)
const ComplexPoint3D(x::NTuple{3, ComplexF64})

A complex 3D point, stored in a StaticArray. ComplexPoint3D = SVector{3, ComplexF64}.

WavePropBase.Point2D — Type

const Point2D
const Point2D(x1, x2)
const Point2D(x::NTuple{2, Float64})

A point in 2D space, stored in a StaticArray. Point2D = SVector{2, Float64}.

WavePropBase.Point3D — Type

const Point3D
const Point3D(x1, x2, x3)
const Point3D(x::NTuple{3, Float64})

A point in 3D space, stored in a StaticArray. Point3D = SVector{3, Float64}.

WavePropBase.ReferenceLine — Type

struct ReferenceLine

Singleton type representing the [0,1] segment.

WavePropBase.ReferenceSquare — Type

struct ReferenceSquare

Singleton type representing the square with vertices (0,0),(0,1),(1,1),(1,0)

WavePropBase.ReferenceTetrahedron — Type

struct ReferenceTetrahedron

Singleton type representing the tetrahedron with vertices (0,0,0),(0,0,1),(0,1,0),(1,0,0)

WavePropBase.ReferenceTriangle — Type

struct ReferenceTriangle

Singleton type representing the triangle with vertices (0,0),(0,1),(1,0)

WavePropBase.abstractmethod — Method

abstractmethod

A method of an abstract type for which concrete subtypes are expected to provide an implementation.

WavePropBase.assert_extension — Function

assert_extension(fname,ext,[msg])

Check that fname is of extension ext. Print the message msg as an assertion error otherwise.

WavePropBase.blockmatrix_to_matrix — Method

blockmatrix_to_matrix(A::Matrix{B}) where {B<:SMatrix}

Convert a Matrix{B}, where B<:SMatrix, to the equivalent Matrix{T}, where T = eltype(B)

WavePropBase.blockvector_to_vector — Method

blockvector_to_vector(A::Vector{B}) where {B<:SVector}

Convert a Vector{B}, where B<:SVector, to the equivalent Vector{T}, where T = eltype(B)

WavePropBase.enable_debug — Method

enable_debug(mname)

Activate debugging messages by setting the environment variable JULIA_DEBUG to mname.

WavePropBase.matrix_to_blockmatrix — Method

matrix_to_blockmatrix(A::Matrix,B)

Convert a Matrix{T} to a Matrix{B}, where B<:Type{SMatrix}. The element type of B must match that of A, and the size of A must be divisible by the size of B along each dimension.

WavePropBase.notimplemented — Method

notimplemented()

Things which should probably be implemented at some point.

WavePropBase.print_threads_info — Method

print_threads_info()

Prints in console the total number of threads.

WavePropBase.svector — Method

svector(f,n)

Just like Base.ntuple, but convert output to an SVector.

WavePropBase.vector_to_blockvector — Method

vector_to_blockvector(A::Vector,B)

Convert a Vector{T} to a Vector{B}, where B<:Type{SVector}. The element type of B must match that of A, and the size of A must be divisible by the size of B along each dimension.